# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CONEXT API",
        "version": "0.3"
    },
    "paths": {
        "/api/abouts": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "Get list of about",
                "operationId": "App\\Http\\Controllers\\AboutController::index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errcode": {
                                            "description": "The response code",
                                            "type": "integer"
                                        },
                                        "errmsg": {
                                            "description": "The response message",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "code": 200,
                                        "data": {
                                            "id": 1,
                                            "content": "<p>In the end of 2017, founders of ST United company decided to invest in coworking space business after long time focusing mainly on offshore software development industry.</p><p>Later, from April 2018, there were 2 new coworking spaces in Danang city , IoT Space and Nomad Space. It has been a new amazing journey for the founders and they figured out that behind this low margin business, coworking space unlike any other office rental facilities, it has a philosophy and contains core values including collaboration, openness, sustainability, accessibility, and community, being a key propellant to why coworking space has become a global trend in business.</p><p>And the idea of Conext is come up with above core values.</p><p>Conext offers a vibrant atmosphere where you can meet others, find partner/cofounder, network, schedule meetups, and most importantly accelerate your works through collaboration with local digital community in Vietnam.</p><p>At Conext, you could connect with startups, digital nomads, freelancers or even investor, incubators…who’s belonging to the community from massive of coworking spaces in the busiest cities in Vietnam.</p><p>Will you join our community, and together we continue writing our story ?</p>",
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "About not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/business-types/{id}": {
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete business-types",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\BusinessTypeController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The business-types that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Business type not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/categories/{id}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The category that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/contact/{id}": {
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete contact",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ContactController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The contact that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Contact not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/events/{id}": {
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete events",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\EventController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The event that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/expertises/{id}": {
            "delete": {
                "tags": [
                    "Expertises"
                ],
                "summary": "Delete expertises",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ExpertiseController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The expertises that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Expertise not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/explore/{id}": {
            "delete": {
                "tags": [
                    "Explore"
                ],
                "summary": "Delete explore",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ExploreController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The explore that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Explore not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/hot-needs/{id}": {
            "delete": {
                "tags": [
                    "Hot-needs"
                ],
                "summary": "Delete hot-needs",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\HotNeedController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The hot-needs that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "HotNeed not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/local-conexts/{id}": {
            "delete": {
                "tags": [
                    "Local-conexts"
                ],
                "summary": "Delete local-conexts",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\LocalConextController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The local-conexts that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Local conext not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete users",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\UserController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The users that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Users not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/vouchers/{id}": {
            "delete": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Delete vouchers",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\VoucherController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The vouchers that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/work-places/{id}": {
            "delete": {
                "tags": [
                    "Work-places"
                ],
                "summary": "Delete work-places",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\WorkPlaceController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The work-places that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Workplace not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}